scripts/data-collection/requirements.txt
mariadb==1.1.10
python-decouple==3.8
GitPython==3.1.43
python-decouple==3.8
ruamel_yaml==0.18.6
pyjson==1.4.1
PyGithub==2.3.0
